requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
faker==25.0.0
openai==1.30.0
httpx==0.25.0
python-dotenv==1.0.1
requests==2.32.0
openenv